feat:新增批量拆箱接口

+++ acc-module/src/main/java/com/ztl/gym/code/service/impl/CodeServiceImpl.java
@@ -546,6 +546,7 @@ public class CodeServiceImpl implements ICodeService {
             throw new CustomException("请输入正确的码！");
         }
     }
+
     @Override
     @DataSource(DataSourceType.SHARDING)
     @Transactional(rollbackFor = Exception.class)
@@ -559,20 +560,25 @@ public class CodeServiceImpl implements ICodeService {
                 code.setpCode(list.get(i));
                 code.setCompanyId(companyId);
                 Code codeEntity = selectCode(code);//查询码产品基本信息
-                if (codeEntity.getStorageRecordId() != null) {
-                    //解除库存(t_product_stock、t_product_stock_flow)
-                    Long inId = codeEntity.getStorageRecordId(); //查询入库表ID
-                    productStockFlowService.unBindProductStockFlowByInId(companyId, inId);
-                    //解除物流明细、解除出入库明细(t_in_code_flow、t_storage_in)
-                    storageInService.unBindStorageInByInId(companyId, inId);
+                if (codeEntity != null) {
+                    try {
+                        if (codeEntity.getStorageRecordId() != null) {
+                            //解除库存(t_product_stock、t_product_stock_flow)
+                            Long inId = codeEntity.getStorageRecordId(); //查询入库表ID
+                            productStockFlowService.unBindProductStockFlowByInId(companyId, inId);
+                            //解除物流明细、解除出入库明细(t_in_code_flow、t_storage_in)
+                            storageInService.unBindStorageInByInId(companyId, inId);
+                        }
+                        if (codeEntity.getCodeAttrId() != null) {
+                            Long attrId = codeEntity.getCodeAttrId();//属性ID
+                            //解除码属性(t_code_attr)
+                            codeAttrService.deleteCodeAttrById(attrId);
+                            //解除绑定关系(t_code)
+                            unBindCodeByAttrId(companyId, attrId);
+                        }
+                    }catch (Exception e){}
                 }
-                if (codeEntity.getCodeAttrId() != null) {
-                    Long attrId = codeEntity.getCodeAttrId();//属性ID
-                    //解除码属性(t_code_attr)
-                    codeAttrService.deleteCodeAttrById(attrId);
-                    //解除绑定关系(t_code)
-                    unBindCodeByAttrId(companyId, attrId);
-                }
+
             }
             return 1;
         } catch (Exception e) {
